move: ContentBox.tsx → customize/ContentBox.tsx
