+++ src/components/Profile.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@chakra-ui/react';
 import { MdPerson } from 'react-icons/md';
-import ContentBox from '@/components/ContentBox';
+import ContentBox from '@/components/customize/ContentBox';
 
 const Profile = () => {
   return (
