+++ src/components/Skill.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@chakra-ui/react';
 import { MdCode } from 'react-icons/md';
-import ContentBox from '@/components/ContentBox';
+import ContentBox from '@/components/customize/ContentBox';
 
 const Skill = () => {
   return (
